misc/HtmlTestSuite/results/test-1145.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
</head>
<body>
    <script>
      test(t =&gt; {
        var i = document.createElement('iframe');
        assert_equals('', i.csp);
        assert_true('csp' in i);
        assert_equals('string', typeof i.csp);
      }, "&lt;iframe&gt; has a 'csp' attibute which is an empty string if undefined.");

      test(t =&gt; {
        var i = document.createElement('iframe');
        i.setAttribute('csp', 123456);
        assert_equals('123456', i.csp);
      }, "&lt;iframe&gt;'s csp attribute is always a string.");

      test(t =&gt; {
        var i = document.createElement('iframe');
        i.csp = 'value';
        assert_equals('value', i.getAttribute('csp'));
      }, "&lt;iframe&gt;'s 'csp content attribute reflects the IDL attribute.");

      test(t =&gt; {
        var i = document.createElement('iframe');
        i.setAttribute('csp', 'value');
        assert_equals('value', i.csp);
      }, "&lt;iframe&gt;'s IDL attribute reflects the DOM attribute.");
  </script>
</body>
</html>
